% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Checks the normality assumption of a variable.}
\usage{
check_normality(
  x,
  test_type = "SW",
  sig_level = 0.05,
  include_graph = TRUE,
  include_interpretation = TRUE
)
}
\arguments{
\item{x}{A single variable.}

\item{test_type}{The type of normality test to be used. Shapiro-Wilk by default}

\item{sig_level}{The significance level for the test. 0.05 by default.}

\item{include_graph}{Whether to include histogram with normal distribution overlay. TRUE by default.}

\item{include_interpretation}{Whether to include interpretation. TRUE by default.}
}
\description{
Checks the normality assumption of a variable.
}
