% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{interpreter}
\alias{interpreter}
\title{Helper function: Creates interpretation of results based on p-value and inputted significance level}
\usage{
interpreter(x, var_name, test_type, sig_level, result)
}
\arguments{
\item{x}{A single variable.}

\item{var_name}{The name of x to be used in the interpretation}

\item{test_type}{The type of normality test to be used. Shapiro-Wilk by default}

\item{sig_level}{The significance level for the normality test}

\item{result}{The output of the normality test used}
}
\value{
Interpretation text as a string
}
\description{
Helper function: Creates interpretation of results based on p-value and inputted significance level
}
