% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{make_histogram}
\alias{make_histogram}
\title{Helper function: Creates histogram of data with a normal distribution overlay}
\usage{
make_histogram(x, var_name)
}
\arguments{
\item{x}{A single variable.}

\item{var_name}{The name of x to be used the histogram title}
}
\description{
Helper function: Creates histogram of data with a normal distribution overlay
}
