% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variance.R
\name{make_interpretation}
\alias{make_interpretation}
\title{Makes an interpretation for check_Variance}
\usage{
make_interpretation(test_type, sig_level, sig)
}
\arguments{
\item{test_type}{The type of test ran}

\item{sig_level}{Significance level of the test}

\item{sig}{Whether or not the result is significant}
}
\value{
A string
}
\description{
Makes an interpretation for check_Variance
}
