% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variance.R
\name{make_plot}
\alias{make_plot}
\title{Makes a residual vs fitted plot for check_variance}
\usage{
make_plot(model)
}
\arguments{
\item{model}{The linear model created by the provided y and x variables}
}
\value{
A plot
}
\description{
Makes a residual vs fitted plot for check_variance
}
