% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linearity.R
\name{make_residual_plot}
\alias{make_residual_plot}
\title{Helper function: Creates the predicted vs. residuals plot}
\usage{
make_residual_plot(preds, residuals)
}
\arguments{
\item{preds}{A single predictor variable.}

\item{residuals}{A numeric residual in a response variable.}
}
\value{
A residual vs. predicted plot
}
\description{
Helper function: Creates the predicted vs. residuals plot
}
