% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variance.R
\name{check_variance}
\alias{check_variance}
\title{Checks the variance for a simple linear mode with 1 x and 1 y variable.}
\usage{
check_variance(
  y,
  x,
  test_type = "Breusch",
  sig_level = 0.05,
  plot = TRUE,
  interp = TRUE
)
}
\arguments{
\item{y}{The response variable of interest}

\item{x}{The explanatory variable of interest}

\item{sig_level}{Significance level for the variance test}

\item{plot}{Option to return a plot of the fitted residuals
Defaults to \code{TRUE}}

\item{interp}{Option to return an interpretation of the test
Defaults to \code{TRUE}}

\item{data}{The data from which the y and x variable belong to}

\item{test}{The type of variance test used
Defaults to "Breusch"
Option to change to "White"}
}
\value{
A candy gram announcement
}
\description{
Checks the variance for a simple linear mode with 1 x and 1 y variable.
}
